Add Hop form, extend use, add volume
Add hop form to hop variety base. Extend use to include whirlpool, hop
stand. Added volume, needed for extract additions, and temperature which
is needed for whirlpool, hop back, and hop stands, to Hop addition
array.

Not sure if volume is added correctly, or if it should be a second unit
type for amount (mass or volume).