Add equipment.json draft
I think I did this right? Please review. 

Tried to limit properties to what is necessary information describing the equipment that does not describe the recipe. Note the lack of boil size, batch size etc which are recipe variables and should not be included in equipment data sets. Lauter Tun added to cover the user cases where sparging occurs in a separate vessel, which is not uncommon for BIAB brewers that sparge. 

UnitTypes, and documentation to be updated next.